refactor(project-management): modularize work run service
Split the 1,755-line work_run_service implementation into modules organized by enqueue, read, dispatch, terminal, cursor, path-lock, and shared persistence responsibilities. Preserve the existing flat public API and all persistence and lifecycle behavior.

Verification: cargo fmt, cargo check --all-targets, focused work_run_service tests, and cargo clippy --all-targets with warnings denied.

Pre-commit hook ran. Total eslint: 0, total circular: 0
